% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{add}
\alias{add}
\title{Arithmetic operations for messydates}
\usage{
add(x, digit, component = "year")
}
\arguments{
\item{x}{A date or list of dates}

\item{digit}{How many units?}

\item{component}{What date component?
By default year.
Options are year, month, or day.}
}
\value{
A messydates list
}
\description{
Arithmetic operations for messydates
}
