% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\alias{on_or_before}
\alias{on_or_after}
\alias{add_approximation}
\alias{add_uncertainty}
\title{Annotates messy dates for uncertainty and/or approximation}
\usage{
on_or_before(x)

on_or_after(x)

add_approximation(x, level = NULL)

add_uncertainty(x, level = NULL)
}
\arguments{
\item{x}{A date vector}

\item{level}{Should the annotation be on the "year", "month", "day",
or month and day ("md")? Optional.
If unspecified annotation will be added to before date.
For month and day ("md") uncertainty or approximation
sign is added before month in date (e.g. 1916-?10-10 or 1916-~10-10).}
}
\value{
A messydt object with annotated date(s)
}
\description{
Some datasets have an arbitrary cut off point
for start and end points, for example,
but these do not refer to the real start or end dates.
Though these are often coded as precise dates.
This collection of functions helps annotate uncertainty and
approximation to dates according to ISO2019E standards.
Innacurate start or end dates can be represented by an affix
indicating "on or before", if used as a prefix (e.g. ..1816-01-01),
or indicating "on or after", if used as a suffix (e.g. 2016-12-31..).
Approximate dates are indicated by adding a \code{~} to year,
month, or day, to estimate values that are
possibly correct (e.g. 2003~-03-03).
Day, month, or year, uncertainty can be indicated by adding a \verb{?}
to a possibly dubious date (e.g. ?1916-10-10) or date
component (e.g. 1916-10-10?).
}
\details{
\code{on_or_before()} annotates uncertain start dates by adding ".." as a prefix
to dates (e.g. \code{..1816-01-01}).

\code{on_or_after()} annotates uncertain end dates by adding ".." as a suffix
to dates (e.g. \code{2016-12-31}).

\code{add_approximation()} annotates approximate dates, or date components,
deemed possibly correct by adding "~" to date (e.g. \code{1916~-01-01})

\code{add_uncertainty()} annotates uncertain dates, or date components,
deemed dubious by adding "?" to date (e.g. \code{1916?-01-01})
}
\examples{
data <- data.frame(Beg = c("1816-01-01", "1916-01-01", "2016-01-01"),
End = c("1816-12-31", "1916-12-31", "2016-12-31"))
dplyr::mutate(data, Beg = ifelse(Beg <= "1816-01-01", on_or_before(Beg), Beg))
dplyr::mutate(data, End = ifelse(End >= "2016-01-01", on_or_after(End), End))
dplyr::mutate(data, Beg = ifelse(Beg == "1916-01-01", add_approximation(Beg), Beg))
dplyr::mutate(data, End = ifelse(End == "1916-12-31", add_uncertainty(End), End))
}
