% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_to.R
\name{as.POSIXlt.messydt}
\alias{as.POSIXlt.messydt}
\title{Coercion from messydate to POSIXlt}
\usage{
\method{as.POSIXlt}{messydt}(x, resolve = c("min", "max", "median", "mean"))
}
\arguments{
\item{x}{a messydt object}

\item{resolve}{How do you want the date ranges to be resolved?
Before coercing dates, one should chose how to resolve ranges
so that dates can be coerced to other date classes.
Choices are "min","max","median", or "mean".}
}
\description{
Coerces from messydt class to POSIXlt class
}
