% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_from.R
\name{as_messydate.Date}
\alias{as_messydate.Date}
\title{Coerce from date class to messydt class}
\usage{
\method{as_messydate}{Date}(x)
}
\arguments{
\item{x}{date class object}
}
\description{
Coerce from date class to messydt class
}
