% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_from.R
\name{as_messydate.POSIXct}
\alias{as_messydate.POSIXct}
\title{Coerce from POSIXct to messydt class}
\usage{
\method{as_messydate}{POSIXct}(x)
}
\arguments{
\item{x}{POSIXct class object}
}
\description{
Coerce from POSIXct to messydt class
}
