% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_from.R
\name{as_messydate.POSIXlt}
\alias{as_messydate.POSIXlt}
\title{Coerce from POSIXlt to messydt class}
\usage{
\method{as_messydate}{POSIXlt}(x)
}
\arguments{
\item{x}{POSIXlt class object}
}
\description{
Coerce from POSIXlt to messydt class
}
