% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_to_messydate.R
\name{as_messydate}
\alias{as_messydate}
\alias{as_messydate.Date}
\alias{as_messydate.POSIXct}
\alias{as_messydate.POSIXlt}
\alias{as_messydate.character}
\title{Coercion to messy dates}
\usage{
as_messydate(x, resequence = FALSE)

\method{as_messydate}{Date}(x, resequence = FALSE)

\method{as_messydate}{POSIXct}(x, resequence = FALSE)

\method{as_messydate}{POSIXlt}(x, resequence = FALSE)

\method{as_messydate}{character}(x, resequence = FALSE)
}
\arguments{
\item{x}{A scalar or vector of a class that can be coerced into a Date,
such as \code{Date}, \code{POSIXct}, \code{POSIXlt}, or character.}

\item{resequence}{Users have the option to choose the
order for ambiguous 6 digit dates (eg. "11-01-12"),
and to expand these dates into precise dates (i.e. YYYY-MM-DD format).
By default FALSE.
If TRUE, it prompts users to select the existing component order of ambiguous
6 digit dates, based on which the date is reordered into YY-MM-DD format
and further completed to YYYY-MM-DD format if they choose to do so.}
}
\value{
A \code{messydt} class object
}
\description{
These functions coerce different data classes into \code{messydt} class
}
\details{
The function can also extract dates from text.
Currently this only works for texts in English.
}
\section{Methods (by class)}{
\itemize{
\item \code{Date}: Coerce from \code{Date} to \code{messydt} class

\item \code{POSIXct}: Coerce from \code{POSIXct} to \code{messydt} class

\item \code{POSIXlt}: Coerce from \code{POSIXlt} to \code{messydt} class

\item \code{character}: Coerce character date objects to \code{messydt} class
}}

\examples{
as_messydate("2021")
as_messydate("2021-02")
as_messydate("2021-02-01")
as_messydate("01-02-2021")
as_messydate("1 February 2021")
as_messydate("First of February, two thousand and twenty-one")
as_messydate("2021-02-01?")
as_messydate("2021-02-01~")
as_messydate("2021-02-01\%")
as_messydate("2021-02-01..2021-02-28")
as_messydate("{2021-02-01,2021-02-28}")
as_messydate(c("-2021", "2021 BC", "-2021-02-01"))
#as_messydate("01-02-21", resequence = TRUE)
}
