% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_battles.R
\docType{data}
\name{battles}
\alias{battles}
\title{Dates of battles in 2001}
\format{
A data frame with 20 rows and 5 variables:
\describe{
\item{Battle}{name of the battle, character}
\item{Date}{date or date range, a mdate class vector}
\item{Parties}{parties to the conflict, character}
\item{US_party}{is the US a party to the battle, numeric}
\item{N_actors}{number of actors to conflict, numeric}
}
}
\usage{
battles
}
\description{
A dataset containing the names and dates of battles in 2001,
according to Wikipedia (https://en.wikipedia.org/wiki/List_of_battles_in_the_21st_century).
}
\examples{
battles$Date
as.Date(battles$Date, FUN = vmin)
}
\keyword{datasets}
