% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mdate.R
\name{class_make}
\alias{class_make}
\alias{make_messydate}
\title{Composes \code{mdate} from multiple variables}
\usage{
make_messydate(..., resequence = FALSE)
}
\arguments{
\item{...}{One (yyyy-mm-dd), two (yyyy-mm-dd, yyyy-mm-dd),
or three (yyyy, mm, dd) variables.}
}
\description{
Composes \code{mdate} from multiple variables
}
\details{
If three date variables are passed to \code{make_messydate()},
function will create a single date (yyyy-mm-dd) from it.
If two date variables are passed to \code{make_messydate()},
function will create a range of dates from it (yyyy-mm-dd..yyyy-mm-dd).
If one date variable is passed to \code{make_messydate()},
function defaults to \code{as_messydate()}.
}
\examples{
make_messydate("2010", "10", "10")
}
