% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_resolve.R
\name{coerce_resolve}
\alias{coerce_resolve}
\alias{vmin}
\alias{vmin.mdate}
\alias{min.mdate}
\alias{vmax}
\alias{vmax.mdate}
\alias{max.mdate}
\alias{random}
\alias{random.mdate}
\title{Resolves messy dates into a single value}
\usage{
vmin(..., na.rm = FALSE)

\method{vmin}{mdate}(..., na.rm = TRUE)

\method{min}{mdate}(..., na.rm = TRUE)

vmax(..., na.rm = FALSE)

\method{vmax}{mdate}(..., na.rm = TRUE)

\method{max}{mdate}(..., na.rm = TRUE)

random(..., size, replace = FALSE, prob = NULL, recursive = FALSE)

\method{random}{mdate}(..., size, replace = FALSE, prob = NULL, recursive = FALSE)
}
\arguments{
\item{...}{a mdate object}

\item{na.rm}{Should NAs be removed? True by default.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights
for obtaining the elements of the vector being sampled.}

\item{recursive}{If recursive = TRUE, then the dates will be resolved
to a single date. If recursive = FALSE, then the dates will be resolved
to a vector the length of the original vector.
By default FALSE.}
}
\value{
A single scalar or vector of dates
}
\description{
This collection of S3 methods 'resolve' messy dates into a single date
according to some explicit bias,
such as returning the minimum or maximum date,
the mean, median, or modal date,
or a random date from among the possible resolutions for each messy date.
If the date is not 'messy' (i.e. has no annotations)
then just that precise date is returned.
This can be useful for various descriptive or inferential projects.
}
\examples{
d <- as_messydate(c("2008-03-25", "?2012-02-27", "2001-01?", "2001~",
  "2001-01-01..2001-02-02", "{2001-01-01,2001-02-02}",
  "{2001-01,2001-02-02}", "2008-XX-31", "-0050-01-01"))
d
vmin(d)
min(d)
vmax(d)
max(d)
random(d)
}
