% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{contract}
\alias{contract}
\title{Contract lists of dates into messy dates}
\usage{
contract(x = list())
}
\arguments{
\item{x}{A list of dates}
}
\value{
A \code{messydt} vector
}
\description{
These functions contract a list of dates into the abbreviated annotation
of messy dates.
}
\examples{
d <- as_messydate(c("2001-01-01", "2001-01", "2001", "2001-01-01..2001-02-02", "{2001-01-01,2001-02-02}", "{2001-01,2001-02-02}"))
e <- expand(d)
tibble::tibble(d,contract(e))
}
