% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_expand.R
\name{expand}
\alias{expand}
\title{Expand Dates}
\usage{
expand(x)
}
\description{
These functions exapand on unspecified dates, ranges and on sets of dates.
Uncertain dates may include several possible dates.
The function "opens" these values to include all the possible dates
contained in uncertain dates.
}
