% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{year}
\alias{month}
\alias{day}
\title{Extracting components from messy dates}
\usage{
year(x)

month(x)

day(x)
}
\description{
Extracting components from messy dates
}
\examples{
year(as_messydate(c("2012-02-03","2012","2012-02")))
month(as_messydate(c("2012-02-03","2012","2012-02")))
day(as_messydate(c("2012-02-03","2012","2012-02")))
}
