% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimr.R
\name{get_skimmers.messydt}
\alias{get_skimmers.messydt}
\title{Adds messydt class \code{{skimr}} template}
\usage{
\method{get_skimmers}{messydt}(column)
}
\arguments{
\item{column}{A messydt object.}
}
\value{
A \code{{skimr}} template for messydt objects.
}
\description{
Defines a \code{{skimr}} template for messydt objects.
The helper functions take a \code{{messydt}} vector as an input and
return a single value to be displayed by the \code{{skimr}} report.
The vector is first resolved yielding a single
statistic per observation before being summarized again by a function
(mean, max, min, etc.) at the vector level.
}
\examples{
d <- tibble::tibble(event = c("Event1", "Event2", "Event3", "Event 4"),
                     messydates = as_messydate(c("2001",
                                                 "2001-01-01..2003-12-30",
                                                 "{2001, 2002, 2003}",
                                                 "33 BC")))
skimr::skim(d)
}
