% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\alias{to_pool}
\title{Imputation for modelling using mdates}
\usage{
impute(.data, times = 1000)

to_pool(imputes)
}
\arguments{
\item{.data}{A dataset with mdate variables.}

\item{times}{Numebr of times for imputation.}

\item{imputes}{Imputed data.}
}
\description{
Imputation for modelling using mdates
}
\examples{
lapply(impute(battles, 3), function(x) mean(x$Date))
#library(psfmi)
#res <- to_pool(impute(battles, 10)) |>
#   dplyr::mutate(Date = as.numeric(Date)) |>
#   psfmi_coxr(Surv(Date, status) ~ US_party + N_actors, impvar = "impute_num")
#res$RR_model_final
}
