% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resequence.R
\name{interleave}
\alias{interleave}
\title{Helper function for interleaving two vectors by position}
\usage{
interleave(vect, pos, elems = NA)
}
\arguments{
\item{vect}{Main vector}

\item{pos}{Positions to be inserted}

\item{elems}{Elements to be inserted at those positions.
By default, these are NAs (missing values).}
}
\value{
A vector the length of the sum of \code{vect}
and \code{pos}.
}
\description{
Insert elements in different positions for vectors
}
