% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{is_sequence}
\alias{is_sequence}
\title{Logical sequence test helper function}
\usage{
is_sequence(x)
}
\arguments{
\item{x}{A date range in messydates format}
}
\value{
TRUE if messydate is sequence or FALSE if not.
}
\description{
Logical sequence test helper function
}
