% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_resolve.R
\name{max.messydt}
\alias{max.messydt}
\title{Resolves for maximum values}
\usage{
\method{max}{messydt}(..., na.rm = TRUE)
}
\arguments{
\item{...}{a messydt object}

\item{na.rm}{Should NAs be removed? True by default.}
}
\description{
Gets maximum values from expanded date ranges,
unspecified dates or sets of dates.
}
