% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{md_duration}
\alias{md_duration}
\title{Get the duration of a date interval}
\usage{
md_duration(x, y, resolve)
}
\arguments{
\item{x}{A messydt object}

\item{y}{A messydt object}

\item{resolve}{How do you want to resolve expanded messy dates
into a single date?
Options are: "min", "max", "mean", "median", "modal", and "random".}
}
\value{
The numerical difference of dates in days
}
\description{
Get the duration of a date interval
}
\examples{
md_duration(as_messydate("2001"), as_messydate("2003"), "min")
md_duration(as_messydate("2001"), as_messydate("2003"), "max")
md_duration(as_messydate("2001"), as_messydate("2003"), "mean")
md_duration(as_messydate("2001"), as_messydate("2003"), "median")
md_duration(as_messydate("2001"), as_messydate("2003"), "modal")
md_duration(as_messydate("2001"), as_messydate("2003"), "random")
}
