% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_resolve.R
\name{mean.messydt}
\alias{mean.messydt}
\title{Resolve for mean values}
\usage{
\method{mean}{messydt}(..., na.rm = TRUE)
}
\arguments{
\item{...}{a messydt object}

\item{na.rm}{Should NAs be removed? True by default.}
}
\description{
Gets mean values from expanded date ranges,
unspecified dates or sets of dates.
}
