% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_resolve.R
\name{median.messydt}
\alias{median.messydt}
\title{Resolve for median values}
\usage{
\method{median}{messydt}(..., na.rm = TRUE)
}
\arguments{
\item{...}{a messydt object}

\item{na.rm}{Should NAs be removed? True by default.}
}
\description{
Gets median values from expanded date ranges,
unspecified dates or sets of dates.
}
