% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_resolve.R
\name{min.messydt}
\alias{min.messydt}
\title{Resolves for minimum values}
\usage{
\method{min}{messydt}(..., na.rm = TRUE)
}
\arguments{
\item{...}{a messydt object}

\item{na.rm}{Should NAs be removed? True by default.}
}
\description{
Gets minimum values from expanded date ranges,
unspecified dates or sets of dates.
}
