% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_resolve.R
\name{modal.messydt}
\alias{modal.messydt}
\title{Resolve for modal values}
\usage{
\method{modal}{messydt}(..., na.rm = TRUE)
}
\arguments{
\item{...}{a messydt object}

\item{na.rm}{Should NAs be removed? True by default.}
}
\description{
Gets modal values from expanded date ranges,
unspecified dates or sets of dates.
}
