% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy_report.R
\name{mreport}
\alias{mreport}
\title{mdate report}
\usage{
mreport(data)
}
\arguments{
\item{data}{A \code{tibble} or a \code{data.frame}.}
}
\value{
\code{mreport()()} returns a data report
of class \code{"mreport"}.
}
\description{
mdate report
}
\examples{
mreport(battles)
}
