% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{operate}
\alias{operate}
\alias{add}
\alias{subtract}
\title{Arithmetic operations for messydates}
\usage{
add(x, digit, component = "year")

subtract(x, digit, component = "year")
}
\arguments{
\item{x}{A date or list of dates}

\item{digit}{How many date units?}

\item{component}{What date component?
By default year.
Other options are month and day.}
}
\value{
A messydates list
}
\description{
These operations allow users to add or subtract dates messydate objects.
Messydate objects include imcomplete or uncertain dates,
ranges of dates, negative dates, and date sets.
}
\details{
\code{add()} adds date units to dates, negative dates,
ranges of dates, and sets of dates.
For ranges of dates, it increases the range by the unit on each side.

\code{subtract()} subtracts date units from dates, negative dates,
ranges of dates, and sets of dates.
For ranges of dates, it reduces the range by the unit on each side.
}
\examples{
d <- as_messydate(c("2008-03-25", "-2012-02-27", "2001-01?", "~2001",
"2001-01-01..2001-02-02", "{2001-01-01,2001-02-02}",
"2008-XX-31", "..2002-02-03", "2001-01-03..", "28 BC"))
tibble::tibble(date = d, add = add(d, 3), subtract = subtract(d, 3))
tibble::tibble(date = d, day = add(d, 1, "day"), month = add(d, 1, "month"),
year = add(d, 1))
tibble::tibble(date = d, day = subtract(d, 1, "day"), month = subtract(d, 1, "month"),
year = subtract(d, 1))
}
