% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{parse_date_strings}
\alias{parse_date_strings}
\title{Helper function to parse character for date operations}
\usage{
parse_date_strings(e2)
}
\arguments{
\item{e2}{Date character}
}
\value{
The equivalent amount of days in numeric format
}
\description{
Helper function to parse character for date operations
}
