% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{+.messydt}
\alias{+.messydt}
\title{#' @include class.R
#' @export
setMethod("+", signature(e1 = "messydt", e2 = "numeric"),
function(e1, e2) add(e1, e2))}
\usage{
\method{+}{messydt}(e1, e2)
}
\arguments{
\item{e1}{element one}

\item{e2}{element two}
}
\description{
#' @include class.R
#' @export
setMethod("-", signature(e1 = "messydt", e2 = "numeric"),
function(e1, e2) subtract(e1, e2))
}
