% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional.R
\name{proportional}
\alias{proportional}
\alias{\%<\%.mdate}
\alias{\%>\%.mdate}
\alias{\%>=\%.mdate}
\alias{\%<=\%.mdate}
\alias{\%><\%.mdate}
\alias{\%>=<\%.mdate}
\title{Proportion of messy dates meeting logical test}
\usage{
e1 %<% e2

e1 %>% e2

e1 %>=% e2

e1 %<=% e2

e1 %><% e2

e1 %>=<% e2
}
\arguments{
\item{x, y}{\code{mdate} or other class objects}
}
\value{
A logical vector the same length as the \code{mdate} passed.
}
\description{
These functions provide various proportional tests for messy date objects.
}
\section{Functions}{
\itemize{
\item \code{ \%<\% }: Tests proportion of dates in the first vector
that precede the minimum in the second vector.

\item \code{ \%>\% }: Tests proportion of dates in the first vector
that follow the maximum in the second vector.
Note that this conflicts with \code{{magrittr}}'s pipe,
and so the base R pipe, \verb{|>} is recommended instead.

\item \code{ \%>=\% }: Tests proportion of dates in the first vector
that follow or are equal to the maximum in the second vector.

\item \code{ \%<=\% }: Tests proportion of dates in the first vector
that precede or are equal to the minimum in the second vector.

\item \code{ \%><\% }: Tests proportion of dates in the first vector
that are between the minimum and maximum dates in the second vector.

\item \code{ \%>=<\% }: Tests proportion of dates in the first vector
that are between the minimum and maximum dates in the second vector, inclusive.

}}
\examples{
  as_messydate("2012-06") < as.Date("2012-06-02")
  as_messydate("2012-06") \%<\% "2012-06-02"
  as_messydate("2012-06") > as.Date("2012-06-02")
  as_messydate("2012-06") \%>\% "2012-06-02"
  as_messydate("2012-06") >= as.Date("2012-06-02")
  as_messydate("2012-06") \%>=\% "2012-06-02"
  as_messydate("2012-06") <= as.Date("2012-06-02")
  as_messydate("2012-06") \%<=\% "2012-06-02"
  as_messydate("2012-06") \%><\% as_messydate("2012-06-15..2012-07-15")
  as_messydate("2012-06") \%>=<\% as_messydate("2012-06-15..2012-07-15")
}
