% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman_population.R
\docType{data}
\name{roman_population}
\alias{roman_population}
\title{Roman Population Sizes dataset}
\format{
A data frame with 4 rows and 3 variables:
\describe{
\item{period}{date range of the Roman Empire}
\item{population_min}{minimum population estimate}
\item{population_max}{maximum population estimate}
}
}
\source{
Scheidel, W. (2008). Roman population size: the logic of the debate.
People, land, and politics: demographic developments and the transformation of Roman Italy, 300, 17-70.
}
\usage{
roman_population
}
\description{
The population sizes in the Roman Empire has been a topic of heated debates among some academics.
Various accounts diverge in how they estimate population size and, therefore,
the conclusions about the Roman population at different points in time (see Scheidel2008).
This dataset contains the high and low estimations for the population of mainland
Italy during the Roman Empire (scheidel2008, p. 31).
}
\keyword{datasets}
