% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{subtract}
\alias{subtract}
\title{Helper function to subtract from messydates}
\usage{
subtract(x, n)
}
\arguments{
\item{x}{A date or list of dates}

\item{n}{What do you want to be added?
Please specify in numeric format (i.e. one day is 1, one year is 365).}
}
\value{
A messydates vector
}
\description{
Helper function to subtract from messydates
}
\details{
\code{subtract()} subtracts date units from dates, negative dates,
ranges of dates, and sets of dates.
For ranges of dates, it subtracts the unit on each side.
}
