% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_date.R
\name{text_to_date}
\alias{text_to_date}
\title{Extract dates from text}
\usage{
text_to_date(v)
}
\arguments{
\item{v}{A character vector}
}
\value{
Dates in messydt format
}
\description{
Sometimes dates can be contained in text,
this function extracts those dates from text and
converts them to messydt objects.
}
\details{
If multiple dates are identified in the same row,
only the first date is returned.
}
\examples{
v <- c("This function was created on the 29 September 2021",
"Today is October 12, 2021", "Tomorrow is 13-10-2021")
text_to_date(v)
}
