% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{add}
\alias{add}
\title{Helper function to add to messydates}
\usage{
add(x, n)
}
\arguments{
\item{x}{A mssydate or list of messydates}

\item{n}{What do you want to be added?
Please specify in numeric format (i.e. one day is 1).}
}
\value{
A messydates vector
}
\description{
Helper function to add to messydates
}
\details{
\code{add()} adds date units to dates, negative dates,
ranges of dates, and sets of dates.
For ranges of dates, it adds the unit on each side.
}
