% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{add}
\alias{add}
\title{Arithmetic operations for messydates}
\usage{
add(x, n)
}
\arguments{
\item{x}{A date or list of dates}

\item{n}{What do you want to be added?
Please specify in numeric format (i.e. one day is 1, one year is 365).}
}
\value{
A messydates vector
}
\description{
These operations allow users to add or subtract dates messydate objects.
Messydate objects include incomplete or uncertain dates,
ranges of dates, negative dates, and date sets.
}
\details{
\code{add()} adds date units to dates, negative dates,
ranges of dates, and sets of dates.
For ranges of dates, it increases the range by the unit on each side.
\code{subtract()} subtracts date units from dates, negative dates,
ranges of dates, and sets of dates.
For ranges of dates, it reduces the range by the unit on each side.
}
\examples{
d <- as_messydate(c("2008-03-25", "-2012-02-27", "2001-01?", "~2001",
"2001-01-01..2001-02-02", "{2001-01-01,2001-02-02}",
"2008-XX-31", "..2002-02-03", "2001-01-03..", "28 BC"))
tibble::tibble(date = d, add = d + 1, subtract = d - 1)
}
