% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_to_messydate.R
\name{as_messydate}
\alias{as_messydate}
\alias{as_messydate.Date}
\alias{as_messydate.POSIXct}
\alias{as_messydate.POSIXlt}
\alias{as_messydate.character}
\title{Coercion to messy dates}
\usage{
as_messydate(x, from_text, interactive)

\method{as_messydate}{Date}(x, from_text, interactive)

\method{as_messydate}{POSIXct}(x, from_text, interactive)

\method{as_messydate}{POSIXlt}(x, from_text, interactive)

\method{as_messydate}{character}(x, from_text = FALSE, interactive = FALSE)
}
\arguments{
\item{x}{A scalar or vector of a class that can be coerced into a Date,
such as \code{Date}, \code{POSIXct}, \code{POSIXlt}, or character.}

\item{from_text}{Would you like to extract dates from text?
By default false.
If TRUE, dates are extracted from text.
If multiple dates are identified, only the first date is returned.
Currently, this only works for texts in English.}

\item{interactive}{Would you like to choose the order for ambiguous dates?
By default FALSE.
If TRUE, it allows users to choose the correct order
for ambiguous 6 digit dates.}
}
\value{
A \code{messydt} class object
}
\description{
These functions coerce different data classes into \code{messydt} class
}
\section{Methods (by class)}{
\itemize{
\item \code{Date}: Coerce from \code{Date} to \code{messydt} class

\item \code{POSIXct}: Coerce from \code{POSIXct} to \code{messydt} class

\item \code{POSIXlt}: Coerce from \code{POSIXlt} to \code{messydt} class

\item \code{character}: Coerce character date objects to \code{messydt} class
}}

\examples{
as_messydate("28 BC")
as_messydate("2021")
as_messydate("2021-02")
as_messydate("2021-02-01")
as_messydate("20-01-2021")
as_messydate("01-20-2021")
as_messydate("20 September 2021")
as_messydate("2021-02-01?")
as_messydate("2021-02-01~")
as_messydate("2021-02-01\%")
as_messydate("2021-02-01..2021-02-28")
as_messydate("{2021-02-01,2021-02-28}")
as_messydate("Two of February of two thousand twenty-two",
from_text = TRUE)
#as_messydate("1-2-21", interactive = TRUE)
}
