% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battles.R
\docType{data}
\name{battles}
\alias{battles}
\title{Dates of battles in 2001}
\format{
A data frame with 19 rows and 2 variables:
\describe{
\item{Battle}{name of the battle, character}
\item{Date}{date or date range, a messydt class vector}
}
}
\usage{
battles
}
\description{
A dataset containing the names and dates of battles in 2001,
according to Wikipedia.
}
\keyword{datasets}
