% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battles.R
\docType{data}
\name{battles}
\alias{battles}
\title{Dates of battles in 2001}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{Battle}{name of the battle, character}
\item{Date}{date or date range, a mdate class vector}
\item{Parties}{parties to the conflict, character}
}
}
\usage{
battles
}
\description{
A dataset containing the names and dates of battles in 2001,
according to Wikipedia (https://en.wikipedia.org/wiki/List_of_battles_in_the_21st_century).
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## -------------------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Missing  |  Missing (\%)  |  Min Value  |  Max Value |
## -------------------------------------------------------------------------------------
## |    Battle     |  character  |     0     |       0       |     NA      |     NA     |
## |     Date      |    mdate    |     0     |       0       | 2001-03-08  | 2002-01-03 |
## |    Parties    |  character  |     0     |       0       |    BD-ID    |     NA     |
## -------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
