% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_resolve.R
\name{coerce_resolve}
\alias{coerce_resolve}
\alias{min.mdate}
\alias{max.mdate}
\alias{median.mdate}
\alias{mean.mdate}
\alias{modal}
\alias{modal.mdate}
\alias{random}
\alias{random.mdate}
\title{Resolves messy dates into a single value}
\usage{
\method{min}{mdate}(..., na.rm = TRUE, recursive = FALSE)

\method{max}{mdate}(..., na.rm = TRUE, recursive = FALSE)

\method{median}{mdate}(..., na.rm = TRUE, recursive = FALSE)

\method{mean}{mdate}(..., trim = 0, na.rm = TRUE, recursive = FALSE)

modal(..., na.rm = FALSE, recursive = FALSE)

\method{modal}{mdate}(..., na.rm = TRUE, recursive = FALSE)

random(..., size, replace = FALSE, prob = NULL, recursive = FALSE)

\method{random}{mdate}(..., size, replace = FALSE, prob = NULL, recursive = FALSE)
}
\arguments{
\item{...}{a mdate object}

\item{na.rm}{Should NAs be removed? True by default.}

\item{recursive}{If recursive = TRUE, then the dates will be resolved
to a single date. If recursive = FALSE, then the dates will be resolved
to a vector the length of the original vector.
By default FALSE.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed
from each end of x before the mean is computed.
Values of trim outside that range are taken as the nearest endpoint.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights
for obtaining the elements of the vector being sampled.}
}
\value{
A single scalar or vector of dates
}
\description{
This collection of S3 methods 'resolve' messy dates into a single date
according to some explicit bias,
such as returning the minimum or maximum date,
the mean, median, or modal date,
or a random date from among the possible resolutions for each messy date.
If the date is not 'messy' (i.e. has no annotations)
then just that precise date is returned.
This can be useful for various descriptive or inferential projects.
}
\examples{
d <- as_messydate(c("2008-03-25", "?2012-02-27", "2001-01?", "2001~",
  "2001-01-01..2001-02-02", "{2001-01-01,2001-02-02}",
  "{2001-01,2001-02-02}", "2008-XX-31", "-0050-01-01"))
d
min(d)
max(d)
median(d)
mean(d)
modal(d)
random(d)
}
