% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\alias{expand.messydt}
\title{Expand messy dates to lists of dates}
\usage{
expand(x)

\method{expand}{messydt}(x)
}
\arguments{
\item{x}{A \code{messydt} object}
}
\value{
Values of all possible dates contained in range

Values of all possible dates contained in range
}
\description{
These functions expand on unspecified dates, ranges and on sets of dates.
Uncertain dates may include several possible dates.
The function "opens" these values to include all the possible dates
contained in uncertain dates.
}
\section{Methods (by class)}{
\itemize{
\item \code{messydt}: Expanding messydates
}}

\examples{
d <- as_messydate(c("2001-01-01", "2001-01", "2001", "2001-01-01..2001-02-02",
        "{2001-01-01,2001-02-02}", "{2001-01,2001-02-02}"))
expand(d)
}
