% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_from_messydate.R
\name{from_messydate}
\alias{from_messydate}
\alias{as.Date.messydt}
\alias{as.POSIXct.messydt}
\alias{as.POSIXlt.messydt}
\title{Coercion from messy dates}
\usage{
\method{as.Date}{messydt}(x, ..., FUN)

\method{as.POSIXct}{messydt}(x, ..., FUN)

\method{as.POSIXlt}{messydt}(x, ..., FUN)
}
\arguments{
\item{x}{A messydt object}

\item{...}{Arguments passed on to the S3 generics.}

\item{FUN}{A function that can be used to resolve expanded messy dates
into a single date.
For example, \code{min()}, \code{max()}, \code{mean()}, \code{median()},
\code{modal()}, and \code{random()}.}
}
\description{
Coerces from messydt class to date class
}
\examples{
as.Date(as_messydate("2012-01"), min)
as.Date(as_messydate("2012-01"), mean)
as.Date(as_messydate("2012-01"), max)
as.Date(as_messydate("2012-01"), median)
as.Date(as_messydate("2012-01"), modal)
as.Date(as_messydate("2012-01"), random)
}
