% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_skimmers.mdate}
\alias{get_skimmers.mdate}
\title{#' @describeIn stats Calculate the mean date of a \code{mdate} vector
#' @export
#' @describeIn stats Calculate the median date of a \code{mdate} vector
#' @export
medianmedian <- function(date) {
median(as.Date(date, median), na.rm = TRUE)
}}
\usage{
\method{get_skimmers}{mdate}(column)
}
\description{
#' @describeIn stats Calculate the mean of the most frequent date of a
#' \code{mdate} vector
#' @export
meanmode <- function(date) {
mean(as.Date(date, modal), na.rm = TRUE)
}
}
