% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{logical}
\alias{logical}
\title{Logical tests on messy dates}
\arguments{
\item{x, y}{\code{mdate} or other class objects}
}
\value{
A logical vector the same length as the \code{mdate} passed.
}
\description{
These functions provide various logical tests for messy date objects.
}
