% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messydate_make.R
\name{make_messydate}
\alias{make_messydate}
\title{Make messy dates from multiple variables}
\usage{
make_messydate(...)
}
\arguments{
\item{...}{One (ymd) or three (yyyy, mm, dd) variables}
}
\value{
A character vector containing the
}
\description{
Transforms multiple date inputs contained in different columns into one.
}
\examples{
make_messydate("2010", "10", "10")
}
