% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy_report.R
\name{messy_report}
\alias{messy_report}
\title{Messy data report}
\usage{
messy_report(data)
}
\arguments{
\item{data}{A \code{tibble} or a \code{data.frame}.}
}
\value{
\code{messy_report()()} returns a data report
of class \code{"messy_report"}.
}
\description{
Messy data report
}
\examples{
messy_report(battles)
}
