% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate_arithmetic.R
\name{operate_arithmetic}
\alias{operate_arithmetic}
\alias{+.mdate}
\alias{-.mdate}
\title{Arithmetic operations for messydates}
\usage{
\method{+}{mdate}(e1, e2)

\method{-}{mdate}(e1, e2)
}
\arguments{
\item{e1}{An \code{mdate} or date object.}

\item{e2}{An \code{mdate}, date, or numeric object. Must be a scalar.}
}
\value{
A messydates vector
}
\description{
These operations allow users to add or subtract dates messydate objects.
Messydate objects include incomplete or uncertain dates,
ranges of dates, negative dates, and date sets.
}
\examples{
\donttest{
d <- as_messydate(c("2008-03-25", "-2012-02-27", "2001-01?", "~2001",
"2001-01-01..2001-02-02", "{2001-01-01,2001-02-02}",
"2008-XX-31", "..2002-02-03", "2001-01-03..", "28 BC"))
dplyr::tibble(date = d, add = d + 1, subtract = d - 1)
dplyr::tibble(date = d, add = d + "1 year", subtract = d - "1 year")
as_messydate("2001-01-01") + as_messydate("2001-01-02..2001-01-04")
as_messydate("2001-01-01") + as_messydate("2001-01-03")
as_messydate("2001-01-01..2001-01-04") - as_messydate("2001-01-02")
#as_messydate("2001-01-01") - as_messydate("2001-01-03")
}
}
