% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\alias{min.messydt}
\alias{max.messydt}
\alias{median.messydt}
\alias{mean.messydt}
\alias{modal.messydt}
\alias{random.messydt}
\title{Resolves messy dates into a single value}
\usage{
\method{min}{messydt}(..., na.rm = TRUE)

\method{max}{messydt}(..., na.rm = TRUE)

\method{median}{messydt}(..., na.rm = TRUE)

\method{mean}{messydt}(..., na.rm = TRUE)

\method{modal}{messydt}(..., na.rm = TRUE)

\method{random}{messydt}(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{...}{a messydt object}

\item{na.rm}{Should NAs be removed? True by default.}
}
\description{
This collection of S3 methods 'resolve' messy dates into a single date
according to some explicit bias,
such as returning the minimum or maximum date,
the mean, median, or modal date,
or a random date from among the possible resolutions for each messy date.
If the date is not 'messy' (i.e. has no annotations)
then just that precise date is returned.
This can be useful for various descriptive or inferential projects.
}
\examples{
d <- as_messydate("2014-01-01..2014-01-31")
d
min(d)
max(d)
mean(d)
median(d)
modal(d)
random(d)
}
