% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\alias{maxmax}
\alias{minmin}
\title{Summary statistic functions}
\usage{
maxmax(date)

minmin(date)
}
\arguments{
\item{date}{A \code{messydt} vector}
}
\value{
A statistic describing the \code{messydt} vector.
}
\description{
\code{messydates} provides extensions to a variety of functions with R's stats
package to simplify creating summaries of data with \code{{skimr}}. All functions
take a \code{messydt} vector as an input and return a single statistic. Note that
individual dates are first resolved (second term in the function name) before
a vector level statistic is computed.
}
\section{Functions}{
\itemize{
\item \code{maxmax}: Calculate the farthest date forward of a \code{messydt} vector

\item \code{minmin}: Calculate the farthest date in the past of a \code{messydt}
vector
}}

\seealso{
\code{\link[=get_skimmers]{get_skimmers()}} for customizing the functions called by \code{\link[=skim]{skim()}}.
}
